Grails 2.4.x support
Made necessary updates to build under grails 2.4.x. This was mainly needed as there
were significant changes made to transactions and testing in Grails 2.2.x and this can
now be leveraged. Plus 2.4.x puts the code up to date with the latest (more or less)
of Grails 2.x before Grails 3.x.